+++ CHANGELOG.md
@@ -4,6 +4,7 @@ dev:
   - use multisign for contributions and proofs with enabled accounts
   - use multisign for signing slot selections with enabled accounts
   - log all obtained attestation data if the majority strategy fails to reach threshold
+  - warn if block proposal gas limit is not the expected value
 
 1.9.2:
   - update go-eth2-wallet-dirk to enable mixed signing thresholds for multisign